+++ mobile/android/android-components/components/feature/search/src/test/java/mozilla/components/feature/search/SearchUseCasesTest.kt
@@ -4,13 +4,17 @@
 
 package mozilla.components.feature.search
 
+import android.content.Context
+import androidx.test.core.app.ApplicationProvider
 import mozilla.components.browser.search.SearchEngine
 import mozilla.components.browser.search.SearchEngineManager
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.support.test.any
+import mozilla.components.support.test.mock
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNotNull
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -23,6 +27,8 @@ import org.robolectric.RuntimeEnvironment
 
 @RunWith(RobolectricTestRunner::class)
 class SearchUseCasesTest {
+    private val context: Context
+        get() = ApplicationProvider.getApplicationContext()
 
     private lateinit var searchEngine: SearchEngine
     private lateinit var searchEngineManager: SearchEngineManager
@@ -34,7 +40,7 @@ class SearchUseCasesTest {
         searchEngine = mock(SearchEngine::class.java)
         searchEngineManager = mock(SearchEngineManager::class.java)
         sessionManager = mock(SessionManager::class.java)
-        useCases = SearchUseCases(RuntimeEnvironment.application, searchEngineManager, sessionManager)
+        useCases = SearchUseCases(context, searchEngineManager, sessionManager)
     }
 
     @Test
@@ -67,5 +73,27 @@ class SearchUseCasesTest {
 
         useCases.defaultSearch.invoke(searchTerms, Session.Source.NEW_TAB)
         verify(engineSession).loadUrl(searchUrl)
+    }
+
+    @Test
+    fun `DefaultSearchUseCase invokes onNoSession if no session is selected`() {
+        var createdSession: Session? = null
+
+        `when`(searchEngine.buildSearchUrl("test")).thenReturn("https://search.example.com")
+        `when`(searchEngineManager.getDefaultSearchEngine(RuntimeEnvironment.application)).thenReturn(searchEngine)
+        `when`(sessionManager.getOrCreateEngineSession(any())).thenReturn(mock())
+
+        var sessionCreatedForUrl: String? = null
+
+        val searchUseCases = SearchUseCases(context, searchEngineManager, sessionManager) { url ->
+            sessionCreatedForUrl = url
+            Session(url).also { createdSession = it }
+        }
+
+        searchUseCases.defaultSearch.invoke("test")
+
+        assertEquals("https://search.example.com", sessionCreatedForUrl)
+        assertNotNull(createdSession)
+        verify(sessionManager).getOrCreateEngineSession(createdSession!!)
     }
 }